% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictools.R
\name{evaluate_candidates}
\alias{evaluate_candidates}
\title{Evaluate expansion candidates}
\usage{
evaluate_candidates(candidates, test)
}
\arguments{
\item{candidates}{an ordered character vector of expansion candidates}

\item{test}{a character vector of (held-out) test candidates}
}
\description{
Evaluate expansion candidates
}
