% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictools.R
\name{expand_wildcards}
\alias{expand_wildcards}
\title{Expand words using wildcard expansion
Default uses 'glob' style wildcards ("finan*"), set type='regex' to use
regular expressions}
\usage{
expand_wildcards(ft_model, words, type = c("glob", "regex"))
}
\arguments{
\item{ft_model}{a fastTextR model}

\item{words}{a character vector of words containing wildcards}
}
\value{
a character vector containing the expanded terms
}
\description{
Expand words using wildcard expansion
Default uses 'glob' style wildcards ("finan*"), set type='regex' to use
regular expressions
}
