% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictools.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{Compute k nearest neighbours from multiple seed words
(by iterating over seed words and taking k words with max similarity)}
\usage{
nearest_neighbors(ft_model, words, k = 1000, k_per_word = k)
}
\arguments{
\item{ft_model}{a fastTextR model}

\item{words}{a character vector of words}

\item{k_per_word}{the number of neighbours to get per seed word}

\item{k_total}{the number of neighbours to return}
}
\value{
ordered list of neighbour words with similarities
}
\description{
Compute k nearest neighbours from multiple seed words
(by iterating over seed words and taking k words with max similarity)
}
