% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictools.R
\name{normalized_vectors}
\alias{normalized_vectors}
\title{Helper function to extract and normalize (ie |row|=1) embeddings}
\usage{
normalized_vectors(ft_model, words)
}
\description{
Helper function to extract and normalize (ie |row|=1) embeddings
}
