% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictools.R
\name{similarity_graph}
\alias{similarity_graph}
\title{Create a similarity graph}
\usage{
similarity_graph(similarities, word_metadata = NULL, threshold = 0.5)
}
\arguments{
\item{similarities}{a data frame of word-word similarities (e.g. output from pairwise_similarity)}

\item{frequences}{an optional data frame of word-level metadata.
The first column should be the word names.
If a 'freq' column is included, it is used for node sizes}
}
\value{
an igraph graph object
}
\description{
Create a similarity graph
}
