% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictools.R
\name{expand_wildcards}
\alias{expand_wildcards}
\title{Expand words using wildcard expansion
Default uses 'glob' style wildcards ("finan*"), set type='regex' to use
regular expressions}
\usage{
expand_wildcards(model_or_dfm, words, type = c("glob", "regex"))
}
\arguments{
\item{model_or_dfm}{a fastTextR model or quanteda dfm}

\item{words}{a character vector of words containing wildcards}
}
\value{
a character vector containing the expanded terms
}
\description{
Expand words using wildcard expansion
Default uses 'glob' style wildcards ("finan*"), set type='regex' to use
regular expressions
}
